removed a few unnecessary code.

+++ 2.JavaCore/src/com/codegym/task/task18/task1821/Solution.java
@@ -5,12 +5,8 @@ Symbol frequency
 
 */
 
-import java.io.BufferedReader;
 import java.io.FileInputStream;
 import java.io.IOException;
-import java.io.InputStreamReader;
-import java.nio.charset.StandardCharsets;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.TreeMap;
 
@@ -30,7 +26,7 @@ public class Solution {
         }
 
         for (Map.Entry<Byte, Integer> entry: byteAndFrequency.entrySet()){
-            byte entryByte = (byte) entry.getKey();
+            byte entryByte = entry.getKey();
             System.out.println((char)entryByte + " " + entry.getValue());
         }
     }
